requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.0

# AI 集成
httpx==0.25.2
openai==1.3.0  # 备用，DeepSeek API 兼容 OpenAI 格式

# 数据处理
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2

# macOS 集成（可选）
# pyobjc-core==10.0
# pyobjc-framework-Cocoa==10.0
# pyobjc-framework-EventKit==10.0
# pyobjc-framework-Contacts==10.0

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0